+++ src/lib/components/table/non-redelegatable/NonRedelegatableTableRow.tsx
@@ -1,8 +1,8 @@
 import { Flex, Grid, Text } from "@chakra-ui/react";
 
-import { TableRow } from "lib/components/table";
-import { ValidatorBadge } from "lib/components/ValidatorBadge";
-import type { NonRedelegatable } from "lib/pages/account-details/types";
+import { ValidatorBadge } from "../../ValidatorBadge";
+import { TableRow } from "../tableComponents";
+import type { NonRedelegatable } from "lib/types";
 import { dateFromNow, formatUTC } from "lib/utils";
 
 interface NonRedelegatableTableRowProps {